+++ header.php
@@ -4,7 +4,9 @@
  *
  * @package    Auberge
  * @copyright  2014 WebMan - Oliver Juhas
- * @version    1.2
+ *
+ * @since    1.0
+ * @version  1.2
  */