+++ js/customizer-preview.js
@@ -5,18 +5,14 @@
  * @copyright  2014 WebMan - Oliver Juhas
  *
  * @since    1.0
- * @version  1.2
+ * @version  1.2.5
  */
 
 
 
 
 
-( function( $ ) {
-
-
-
-	"use strict";
+jQuery( function() {
 
 
 
@@ -27,7 +23,7 @@
 		wp.customize( 'blogname', function( value ) {
 
 			value.bind( function( to ) {
-				$( '.site-title a' ).text( to );
+				jQuery( '.site-title a' ).text( to );
 			} );
 
 		} ); // /blogname
@@ -35,11 +31,11 @@
 		wp.customize( 'blogdescription', function( value ) {
 
 			value.bind( function( to ) {
-				$( '.site-description, .site-banner-header .highlight' ).text( to );
+				jQuery( '.site-description, .site-banner-header .highlight' ).text( to );
 			} );
 
 		} ); // /blogdescription
 
 
 
-} )( jQuery );
+} );